Allow custom alignment for symbols
Previously we just forced the ELF class's word size as the symbol
alignment, which allowed us to cheat and just always pad at the end of
symbol values up to the next word boundary.

Sometimes we need more specific alignment though, and allowing that means
we need to adopt a more conventional strategy of dealing with the
alignment padding _before_ writing a symbol, so that we're putting the
start position at a suitable offset.

We also need to make sure that we write the .rodata section with the
greatest alignment of all of the symbols, so that the most constrained of
the symbols will be at its needed alignment.