Allow choosing a different section name (other than .rodata)
The .rodata section is a good default choice for most cases, but some
target platforms want large data in a different section so that e.g. it
can be loaded into some memory that is slower but more plentiful than
where .rodata ends up, or other similar linker script shenanigans.